+++ frontend/src/routes/map/utils/map.ts
@@ -3,3 +3,4 @@ export * from '$routes/map/utils/map/longitude';
 export * from '$routes/map/utils/map/scale';
 export * from '$routes/map/utils/map/style-json';
 export * from '$routes/map/utils/map/tile';
+export * from '$routes/map/utils/map/tile-coordinate';
